move overlays code to headers
