+++ rpcs3/Emu/RSX/Overlays/Shaders/shader_loading_dialog_native.cpp
@@ -1,5 +1,6 @@
 ﻿#include "stdafx.h"
 #include "shader_loading_dialog_native.h"
+#include "../overlay_message_dialog.h"
 
 namespace rsx
 {
