+++ rpcs3/Emu/RSX/Overlays/overlays.cpp
@@ -1,5 +1,6 @@
 ﻿#include "stdafx.h"
 #include "overlays.h"
+#include "overlay_message_dialog.h"
 #include "../GSRender.h"
 #include "Input/pad_thread.h"
