+++ rpcs3/rpcs3qt/trophy_notification_helper.cpp
@@ -2,7 +2,7 @@
 #include "trophy_notification_frame.h"
 
 #include "../Emu/System.h"
-#include "../Emu/RSX/Overlays/overlays.h"
+#include "../Emu/RSX/Overlays/overlay_trophy_notification.h"
 
 s32 trophy_notification_helper::ShowTrophyNotification(const SceNpTrophyDetails& trophy, const std::vector<uchar>& trophy_icon_buffer)
 {
